requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
pydantic-settings>=2.4,<3.0
PyYAML>=6.0
requests>=2.32
beautifulsoup4>=4.12
psutil>=6.0
pytest>=8.3
httpx>=0.27
flake8>=7.1
isort>=5.13
pywin32>=306; platform_system=="Windows"
